// src/app/client-view/models/contact.model.ts
export class Contact {
    id: string;
    fullName: string;
    firstName: string;
    lastName: string;
    gender: number;
    profileImageUrl: string;
    birthday: string;
    streetAddressLine1: string;
    streetAddressLine2: string;
    streetAddressCity: string;
    streetAddressStateOrProvince: string;
    streetAddressPostalCode: string;
    homePhone: string;
    mobilePhone: string;
    email: string;
    maritalStatus: number;
    age: number;
    smoker: boolean;
    height: number;
    weight: number;
    employer: string;
    occupation: string;
    grossSalary: number;
    bmi: number;

    superannuationRate: number;
    taxFileNumber: string;
    currentIncome: number;
    retirementAge: number;
    retirementIncome: number;

    ageGoalCompletePercentage: number;
    incomeGoalCompletePercentage: number;

    houseHoldId: string;
    isDisplayedInUi: boolean;
    indexInList: number;
    role: string;
    onTrack: number;
}